test: fix comparison of data
I think this was actually a harmless problem before, too.
It seems `==` of xarray datasets returns a bool, not an array, so the `np.all`s were redundant.
Also, dataset `==` must have included dimension names. I broke the equality of dimension names,
but not the equality of the data itself, as these tests now show.